feat: create a Picker component for option selection
